Moved icons to their own folder.
