+++ src/assets/icons/PineappleRed.png
