+++ src/components/Footer/Footer.js
@@ -1,6 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import MySHC from "../../assets/images/MySHC.png";
+import MySHC from "../../assets/icons/MySHC.png";
+import FacebookIcon from "../../assets/icons/Facebook.png";
 import "./footer.scss";
 
 const Footer = () => {
@@ -10,6 +11,9 @@ const Footer = () => {
         <img src={MySHC} title={"MySHC Logo"} alt={"MySHC Logo"} className={"mySHCImage"} />
         <Link to={"/login"} className={"link"}>
           Login to mySHC
+        </Link>
+        <Link to={"https://facebook.com/bsushc"}>
+          <img className={"socialIcon"} src={FacebookIcon} title={"Facebook Icon"} alt={"Facebook Icon"} />
         </Link>
       </article>
     </footer>
