+++ src/components/Footer/footer.scss
@@ -35,8 +35,7 @@ footer {
     border-radius: 1rem;
     padding: 0.5rem 1.5rem 0.5rem 1.5rem;
     width: 8rem;
-    margin-left: auto;
-    margin-right: auto;
+    margin: 1rem auto 1rem auto;
     color: $common-white;
     text-decoration: none;
   }
